+++ test/integration/GCClassic/integrationTestCompile.sh
@@ -32,6 +32,13 @@
 #------------------------------------------------------------------------------
 #BOC
 
+#=============================================================================
+# Arguments
+#=============================================================================
+
+# Run a short integration test?
+quick="${1}"
+
 #============================================================================
 # Global variable and function definitions
 #============================================================================
@@ -104,8 +111,15 @@ fi
 # Sanity check: Max out the OMP_STACKSIZE if it is not set
 [[ "X${OMP_STACKSIZE}" == "X" ]] && export OMP_STACKSIZE=500m
 
+# Only create necessary executables if $quick is "yes"
+if [[ "x${quick}" == "xyes" ]]; then
+    EXE_LIST=("default" "carbon")
+else
+    EXE_LIST=$EXE_GCC_BUILD_LIST
+fi
+
 # Count the number of tests to be done
-numTests=${#EXE_GCC_BUILD_LIST[@]}
+numTests=${#EXE_LIST[@]}
 
 #============================================================================
 # Initialize results logfile
@@ -151,7 +165,7 @@ let failed=0
 let remain=${numTests}
 
 # Loop over build directories
-for dir in ${EXE_GCC_BUILD_LIST[@]}; do
+for dir in ${EXE_LIST[@]}; do
 
     # Define build directory
     thisBuildDir="${buildDir}/${dir}"